+++ src/components/calcite-card/resources.ts
@@ -1,0 +1,16 @@
+export const CSS = {
+  container: "container",
+  header: "header",
+  footer: "footer",
+  title: "title",
+  subtitle: "subtitle",
+  thumbnailWrapper: "thumbnail-wrapper"
+};
+
+export const enum SLOTS {
+  thumbnail = "thumbnail",
+  title = "title",
+  subtitle = "subtitle",
+  footerLeading = "footer-leading",
+  footerTrailing = "footer-trailing"
+}